% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{download_gpdd}
\alias{download_gpdd}
\title{Download Data from GPDD.}
\usage{
download_gpdd(dataset_name)
}
\arguments{
\item{dataset_name}{A string of the name of the dataset.}
}
\value{
A dataset in tibble format.
}
\description{
Download Data from GPDD.
}
\examples{
download_gpdd('data')
download_gpdd('timeperiod')
}
\author{
Zihao Li, \email{zihao0104@berkeley.edu}
}
